💄[style] 프로젝트 생성 모달 변경 #134
